Move floating label input to root components
